requirements.txt
langchain
fastapi
uvicorn
langchain-google-genai
google-generativeai
python-dotenv
streamlit